(custom-serialize, custom-deserialize, define-custom-serializer, define-custom-deserializer): functions to do deep conversions across (type, value) pairs
(custom-set-typed-variable): similar to custom-set-variable, but with
an extra arg, the type of the variable. This will slowly replace c-s-v
in `custom' files. The type allows the conversion to work even if the
variable has not yet been defined

(variable-type): new function

(custom-set-color, custom-set-font): removed, just conversion functions
now. Compatibility functions in compat.jl