Ignore grab-induced enter-notify events.
Turns out the focus modes were already ignoring grab-induced leave-notify
events.  Ignoring enter-notify events is a bit of a double-edged sword.

The problem that this solves is when you press Alt F2 in KDE, the Run
Command popup appears and gets focused because you have configured it so,
you start typing, the completion list pops up and takes focus, then you
type a bit further and the completion list disappears.  Focus used to
revert to the window the pointer was in, which usually was not the Run
Command popup.  transient-unmap-window tried hard to get it right, then
the grab-induced enter-notify brought focus to the wrong window anyway.

The problem that this causes is when you open the KDE main menu by
clicking the K icon in the panel, move the mouse into some application
window and click to dismiss the open menu, focus reverts to the panel.
You have to move the pointer out of the application window and back in
to get it focused.

I am talking about the enter-exit and enter-only focus modes of course.
The click focus mode ignores enter-notify events anyway.