New focus mode operation focus-revert.
This replaces the logic in transient-unmap-window that was used to
assign focus when the currently focused window closes.  Passing focus
to the transient parent of the closing window still takes precedence,
but if that is not an option, the decision is now delegated to the
global focus mode.

For now, I am trying to retain the old functionality.

Conflicts:

	lisp/sawfish/wm/focus.jl