apply patch by Thadeu Lima de Souza Cascardo
This is a patch for Debian bug #406559, and possibly #403100, too. It seems these are ultimately caused by Xlib "helpfully" converting 32 bit messages from the server to the client's native long type, which is bad for sawfish since rep does distinguish these.

A longer description can be found in the patch's author's message in http://bugs.debian.org/406559


svn path=/trunk/; revision=4208