Net wm properties by Rodrigo Gallardo
This patch makes sawfish use the _NET_WM_NAME properties for window names, as
mandated by the EWMH spec
(http://standards.freedesktop.org/wm-spec/wm-spec-1.3.html)

This patch is a bit more invasive that it seems requiered, because, to make it
work properly in my system I had to make sawfish assume all strings are UTF8
encoded (see the pieces of the patch touching fonts.c)

Please test it and tell me what breaks. If there are no objections to the
substance of the patch, I'd still like some feedback on what to do with text
rendering with "normal" xlfd fonts. 


GSR-FR comment:
Probably the sanest solution, just next release will have to be marked with a
huge "Reencode strings in utf-8".
     ^^^^^^^^^^^^^^^^^^^^^^^^^^^


svn path=/trunk/; revision=4211