patch by Harald van Dijk and Timo Korvola, "Titlebar updates"
When using a theme which updates itself as window titles change, such as the
default theme Crux, the updates happen at the wrong time, and the title bar
ends up messed up. When the netwm patch is applied, you can easily see this by
normal browsing in Mozilla Firefox, or editing a document in editors that
update the title to include "*" or "modified" when you start typing. This
happens because the themes look for updates to the WM_NAME property, but
sawfish now reads the _NET_WM_NAME property as well, so themes should check for
updates to that too. This patch updates the default themes, and to prevent
custom themes (not distributed with sawfish itself) from breaking, treats
(call-after-property-changed 'WM_NAME ...) as a special case that should also
check for _NEW_WM_NAME.


svn path=/trunk/; revision=4221