New grammar for 'add-window-matcher' and 'remove-window-matcher'. Now it is possible to specify both window's name and class (and others). They now take two arguments. Old usage is still supported. Doc is done.
The old behavior was buggy, and even if the old grammar is used,
it is amended.
 Suppose you've already got rules for '((WM_NAME "foo") (WM_CLASS "class"))
and '((WM_NAME "bar") (WM_CLASS "class")).
Now if you you (add-window-matcher 'WM_CLASS "class" actions...),
then it used to be added to *one of* the above two, and it was unpredictable
which would be! This is fixed.