Fix of a semi-bug. (defgroup stacking) is moved to custom.jl from stacking. Without this, adding 'tab' to open clause of wm.frames breaks the startup.
Reason: In lisp/sawfish/wm.jl, wm.stacking is loaded. But long chain of
'open' starting from wm.stacking reaches wm.util.stacking, which has
 (defcustom foo :group (misc stacking))
But the definition of the group 'stacking' is in body of wm.stacking
which is not yet evaluated! But body eval has to come after 'open', so
it can't be fixed by respecting modularity (of librep).

Of course it can be solved by fixing defcustom & defgroup, but it's
beyond us now. (And isn't it the same reason that Harper decided to
centralize many defgroups into wm.custom?)