move the focus ONCE after processing key command. If sent, schedule Ungrabbing.
Too bad, only now (after years) I realized why the Input-Focus setting is
postponed (in the C code).  I always thought it was for optimization. I realize,
that the Lisp code can set it (possibly non intentionally, just by using hi
level functions).  I calibrated my explicit AllowEvents & (especially)
XUngrabKeyboard to be directed into the right (focused) window.  .... But now, I
connected these 2 facts: that you can ungrab just once. so there is no point in
delegating InputFocus to a window, when we do not want any input from
there (since we still keep a grab, and have not finished doing the command).
Hence it is reasonable to direct the Focus & ungrab after the command (bound the
the pressed key) is all executed.  Now the only trouble with this, is when we
deletegate the Focus setting to the client: by sending wm-take-focus: then we
have to Ungrab when the client chose its window (we receive FocusIn).
+  remove ungrabbing on_idle