Focus-mode `enter-exit' bugfix; it used to react to `pointer-out'. It caused unwanted focus loss.
At least two cases are known. One is when a new instance of gtk's drop down menu appears, and the pointer has been located in the menu's inside area. The other is firefox' urlbar autocompletion; when the cursor lies slightly below the urlbar, and autocompletion appears, it flickers and not few keypresses are lost.

Both were terribly inconvienient.